# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Minecraft Server Monitor API",
    "version": "2.0.0",
    "description": "OpenAPI specification for the Minecraft Server Monitor API used by Open WebUI Tool Server."
  },
  "servers": [
    { "url": "https://gameserver.petedillo.com", "description": "Production" }
  ],
...
  "paths": {
...
    "/v2/api/tp": {
      "post": {
        "tags": ["admin"],
        "operationId": "teleport",
        "summary": "Teleport player to coords or another player",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TeleportRequest" },
              "examples": {
                "toPlayer": {
                  "summary": "Teleport Steve to Alex",
                  "value": { "from": "Steve", "to": "Alex" }
                },
                "toCoords": {
                  "summary": "Teleport Steve to coordinates",
                  "value": { "from": "Steve", "coords": { "x": 100, "y": 70, "z": 200 } }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OkResponse" } } } } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TeleportRequest": {
        "type": "object",
        "properties": {
          "from": { "type": "string" },
          "to": { "type": "string" },
          "coords": {
            "type": "object",
            "properties": {
              "x": { "type": "integer" },
              "y": { "type": "integer" },
              "z": { "type": "integer" }
            },
            "required": ["x", "y", "z"]
          }
        },
        "required": ["from"],
        "oneOf": [
          { "required": ["to"] },
          { "required": ["coords"] }
        ]
      },
...
      "OkResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
...
    }
  }
}